fix: Resolve duplicate sidebar links by renaming pages directory
Streamlit's native multipage feature automatically discovers and adds
pages from a directory named `pages`. This was conflicting with the
manual page loading implemented with `streamlit_option_menu` in
`app_main.py`, causing sidebar links to appear twice.

This commit resolves the issue by renaming the `pages` directory to
`app_pages`. This prevents Streamlit's automatic page discovery,
leaving the manually configured menu as the single source of truth
for navigation. The import statements in `app_main.py` have been
updated to reflect the new directory name.
